Class File Separated
Moved all classes into separate files, so found an error in the FinishDrink function in Actions, where drink size was hard coded to 10
